PRO-C51: DESIGN ORIGAMI WEBSITE - 1
THIS IS MY PROJECT OF CLASS 51. IN THIS CLASS I HAVE LARNT ABOUT THE COLORS USING JAVASCRIPT. HERE, IT IS MY HOMEWORK. IN THIS, Michelle has an origami studio and has made a website for it. She wants to make the website more user-friendly by adding a background color, margin, and padding.